pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-file-manager"
version = "2.0.0"
description = "Enhanced cross-platform audio file manager with OS abstraction, legacy compatibility, and advanced features"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "A.A."},
]
keywords = ["audio", "recording", "playback", "cross-platform", "alsa", "sounddevice"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Platform-specific dependencies are handled in setup.py
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
]
all = [
    "pyalsaaudio>=0.8.0",
    "sounddevice>=0.4.0",
    "numpy>=1.19.0",
]

[project.scripts]
audio-manager-demo = "example_enhanced_manager:main"
audio-record-demo = "enhanced_record_example:main"

[project.urls]
Homepage = "https://github.com/tx3m/audio_files_manager"
Repository = "https://github.com/tx3m/audio_files_manager.git"
Issues = "https://github.com/tx3m/audio_files_manager/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["audio_file_manager*"]

[tool.black]
line-length = 100
target-version = ['py37']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
